Several changes and cleanup
- Moved DefyX conf into RandomX
- Completed Panthera Integration
- Removed obsolete files
