+++ src/crypto/common/Algorithm.h
@@ -72,7 +72,7 @@ public:
         AR2_CHUKWA,    // "argon2/chukwa"    Argon2id (Chukwa).
         AR2_WRKZ,      // "argon2/wrkz"      Argon2id (WRKZ)
         DEFYX,         // "defyx"            DefyX (Scala).	
-        RX_XLA,        // "rx/xla"           DefyX v2 AKA Panthera (Scala).			
+        RX_XLA,        // "rx/xla"           Panthera (Scala).			
         MAX
     };
