+++ src/crypto/randomx/defyx/insecure_memzero.h
