+++ src/crypto/randomx/defyx/yescrypt-platform.c
