+++ src/crypto/randomx/defyx/yescrypt-simd.c
